refactor module and controller. add additional route
