+++ src/app.controller.ts
@@ -1,0 +1,16 @@
+import { Controller,Get } from "@nestjs/common";
+
+@Controller('/app')
+class AppController {
+  @Get()
+  getRootRoute() {
+    return 'Hello World';
+  }
+
+  @Get('/bye')
+  getByeThere(){
+    return "Ok Bye!!!"
+  }
+}
+
+export default AppController
