+++ src/main.ts
@@ -1,18 +1,5 @@
-import { Controller, Module, Get } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
-
-@Controller()
-class AppController {
-  @Get()
-  getRootRoute() {
-    return 'Hello World';
-  }
-}
-
-@Module({
-  controllers: [AppController],
-})
-class AppModule {}
+import AppModule from './app.module';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
